+++ src/ui/main/mod.rs
@@ -14,6 +14,7 @@ mod update_patch;
 mod download_wine;
 mod create_prefix;
 mod download_diff;
+mod disable_telemetry;
 mod launch;
 
 use anime_launcher_sdk::components::loader::ComponentsLoader;
@@ -362,6 +363,8 @@ impl SimpleComponent for App {
                                                 Some(LauncherState::PatchNotInstalled) |
                                                 Some(LauncherState::PatchUpdateAvailable) => "document-save-symbolic",
 
+                                                Some(LauncherState::TelemetryNotDisabled) => "security-high-symbolic",
+
                                                 Some(LauncherState::WineNotInstalled) |
                                                 Some(LauncherState::PrefixNotExists) => "document-save-symbolic",
 
@@ -385,6 +388,8 @@ impl SimpleComponent for App {
 
                                                 Some(LauncherState::PatchBroken) => tr("patch-broken"),
                                                 Some(LauncherState::PatchUnsafe) => tr("patch-unsafe"),
+
+                                                Some(LauncherState::TelemetryNotDisabled) => tr("disable-telemetry"),
 
                                                 Some(LauncherState::WineNotInstalled) => tr("download-wine"),
                                                 Some(LauncherState::PrefixNotExists)  => tr("create-prefix"),
@@ -951,6 +956,8 @@ impl SimpleComponent for App {
 
                     LauncherState::PatchNotInstalled |
                     LauncherState::PatchUpdateAvailable => update_patch::update_patch(sender, self.progress_bar.sender().to_owned()),
+
+                    LauncherState::TelemetryNotDisabled => disable_telemetry::disable_telemetry(sender),
 
                     LauncherState::WineNotInstalled => download_wine::download_wine(sender, self.progress_bar.sender().to_owned()),
                     LauncherState::PrefixNotExists => create_prefix::create_prefix(sender),